+++ R/create_cache_file.R
@@ -22,9 +22,11 @@ create_cache_file <- function(pipeline_inventory = NULL,
                               compress           = 100,
                               verbose            = getOption("pipdm.verbose"),
                               force              = FALSE,
-                              cpi_dt             = NULL,
-                              ppp_dt             = NULL,
-                              pfw_dt             = NULL) {
+                              cpi_dt             = pipload::pip_load_aux("cpi"),
+                              ppp_dt             = pipload::pip_load_aux("ppp"),
+                              pfw_dt             = pipload::pip_load_aux("pfw"),
+                              pop_dt             = pipload::pip_load_aux("pop")
+                              ) {
 
 
   # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -142,7 +144,7 @@ create_cache_file <- function(pipeline_inventory = NULL,
       processed_data = "No data processed",
       data_available = crr
     )))
-  }
+  }  # end of `nrow(new_svy_ids) == 0`
 
   #--------- Process data: Load, clean, and save ---------
   if (verbose) {
@@ -171,7 +173,8 @@ create_cache_file <- function(pipeline_inventory = NULL,
         compress      = 100,
         cpi_dt        = cpi_dt,
         ppp_dt        = ppp_dt,
-        pfw_dt        = pfw_dt
+        pfw_dt        = pfw_dt,
+        pop_dt        = pop_dt
       )
     }
   )
@@ -180,9 +183,9 @@ create_cache_file <- function(pipeline_inventory = NULL,
 
   crr_status <- pip_update_cache_inventory(
     pipeline_inventory = pipeline_inventory,
-    pip_data_dir = pip_data_dir,
-    cache_svy_dir = cache_svy_dir,
-    tool = tool
+    pip_data_dir       = pip_data_dir,
+    cache_svy_dir      = cache_svy_dir,
+    tool               = tool
   )
 
   if (verbose) {