+++ NEWS.md
@@ -1,3 +1,9 @@
+# pipdm (development version)
+* Add function `db_create_pipeline_inventory`
+* add function `delete_old_output_file`
+* Add function `adjust_population`
+
+
 # pipdm 0.0.7
 * remove deprecated files from cache inventory
 * Allow the creation dist  statistics for aggregate group data with only one area, like ARG 1980